requirements.txt
gspread
gdown==5.1.0
google-api-python-client
google-auth
python-dotenv
requests
google-generativeai
streamlit
dnspython
cloudscraper
extra-streamlit-components
google-cloud-storage
gcsfs
supabase
gdown
duckdb>=0.9.0
peopledatalabs
fastapi
uvicorn
pandas
pydantic>=2.0.0
pydantic-settings>=2.0.0
msal
